cache: remove extra pins before adding them back
If directory caching is enabled for .opam, the previous process for
adding pins will fail due to duplicate names, e.g., extra0, extra1, etc.
Removing all pins except the default before attempting to add them
avoids this problem.

For those that are not using directory caching or are not using
EXTRA_REMOTES, the behavior should be unchanged.