Support BASE_REMOTE in .travis-docker.sh
Similar to BASE_REMOTE in .travis-ocaml.sh

This is useful when wanting to test alternative opam repositories that have
packages that haven't made it yet to main opam repository.
EXTRA_REMOTES can be used here but that uses the repo in addition to main one,
with BASE_REMOTE you can test a particular opam remote in isolation.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>